+++ src/icon/chevronIcon.tsx
@@ -2,13 +2,18 @@ import React from 'react'
 
 import { BaseIcon, BaseIconDefaultProps, Icon } from '../_utils/icon'
 
+export enum ChevronIconDirections {
+  DOWN = 'DOWN',
+  UP = 'UP',
+  LEFT = 'LEFT',
+}
+
 export type ChevronIconProps = Icon &
   Readonly<{
-    down?: boolean
-    left?: boolean
+    direction?: ChevronIconDirections
   }>
 
-export const ChevronIcon = ({ down, left, ...props }: ChevronIconProps) => (
+export const ChevronIcon = ({ direction, ...props }: ChevronIconProps) => (
   <BaseIcon {...props}>
     <polyline
       fill="none"
@@ -18,8 +23,9 @@ export const ChevronIcon = ({ down, left, ...props }: ChevronIconProps) => (
       strokeLinejoin="round"
       strokeMiterlimit="10"
       points="9 18 15 12 9 6"
-      {...(down && { transform: 'rotate(90 12 12)' })}
-      {...(left && { transform: 'rotate(180 12 12)' })}
+      {...(direction === ChevronIconDirections.DOWN && { transform: 'rotate(90 12 12)' })}
+      {...(direction === ChevronIconDirections.LEFT && { transform: 'rotate(180 12 12)' })}
+      {...(direction === ChevronIconDirections.UP && { transform: 'rotate(-90 12 12)' })}
     />
   </BaseIcon>
 )