+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 - **[FIX]** Fix broken docs for dividers
 [...]
 
+# v52.0.0 (03/03/2021)
+
+- **[BREAKING CHANGE]** Added enum `direction` prop on `ChevronIcon`
+
+# v51.0.0 (02/03/2021)
+
+- **[BREAKING CHANGE]** rename `HighlightSection` props
+- **[BREAKING CHANGE]** Removed `deprecatedExtraFooter` prop on `ColumnedContentSection`
+- **[UPDATE]** Refactor `ColumnedContentSection` following a compound component approach
+- **[FIX]** Fix top link position on `ColumnedContentSection` and grouping it with the title
+- **[NEW]** Add `FaqSection` component
+
 # v50.0.0 (02/03/2021)
 
 - **[BREAKING CHANGE]** Move `Drawer`, `DropdownButton`, `HamburgerButton` and `Menu` to `TopBar`
@@ -25,7 +37,6 @@
 - **[UPDATE]** Horizontal Normalization on `SearchRecap`
 - **[BREAKING CHANGE]** Removed `href` prop on `UneditableTextField` and Horizontal Normalization
 - **[NEW]** Add `SpecialOffer` component
-
 
 # v48.1.0 (24/02/2021)